Industrial-and-Automotive-Paint-Booths-Market.md
Paint Booths, as the name suggests, is to spray paint and color the filtered spray paint objects in a room. Also known as the paint booth. The spray booth is an equipment that provides a special environment for painting operations, which can meet the requirements of painting operations for temperature, humidity, illuminance, and air cleanliness; it can limit the paint mist and organic waste gas generated during painting operations and discharge after treatment, which is environmentally friendly -Type coating equipment.