Global-Perishable-Goods-Transportation-Market.md
Perishable Goods Transportation involves the transportation of temperature-sensitive goods. It is an integrated component of cold chain logistics and helps in the management and control of the flow of the cold chain supply process. Perishable goods transportation is an integrated component of cold chain logistics and helps in the management and control of the flow of the cold chain supply process.